+++ Craftsman/Models/AddEntityTemplate.cs
@@ -2,6 +2,8 @@
 
 namespace Craftsman.Models
 {
+    using System.Linq;
+
     /// <summary>
     /// This is the complete object representation of the API that we will read in from our input file and scaffold out the necessary files
     /// </summary>
@@ -32,11 +34,12 @@ namespace Craftsman.Models
         /// <summary>
         /// Calculation to determine whether or not authentication is added to the project
         /// </summary>
-        public bool AddJwtAuthentication
-        {
-            get => AuthorizationSettings
-                .Policies
+        public bool AddJwtAuthentication =>
+            Entities
+                .Select(e => e.Features
+                    .Select(f => f.Policies)
+                    .ToList())
+                .ToList()
                 .Count > 0;
-        }
     }
 }